requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
langchain-community
langchain-huggingface
pandas
numpy
matplotlib
Pillow
opencv-python
weasyprint
